+++ test/dummy/app/models/invoice.rb
@@ -1,3 +1,4 @@
 class Invoice < ApplicationRecord
   serialize :data
+  include ApplicationRecordLogger
 end