+++ lib/application_record_logger/service.rb
@@ -1,6 +1,14 @@
 module ApplicationRecordLogger
   class Service
-    attr_accessor :user, :user_id, :config, :action, :record, :record_id, :record_type
+    include ActionParser
+    attr_accessor :user, :user_id, :config, :action
+
+    def initialize(**kwargs, &block)
+      kwargs.each do |key, value|
+        send("#{key}=", value) if respond_to? "#{key}="
+      end
+      yield self if block_given?
+    end
 
     def self.call(**kwargs, &block)
       new(**kwargs, &block).call
@@ -11,19 +19,39 @@ module ApplicationRecordLogger
     end
 
     def logs
-      ::ApplicationRecordLog.where(record: record)
+      ::ApplicationRecordLog.where(record_id: record_id, record_type: record_type)
+    end
+
+    def action=(value)
+      @action = parse_action(value)
     end
 
     def record
-      @record ||= record_klass.find(record_id)
+      @record ||= find_record
+    end
+
+    def record=(rec)
+      @record_id = rec.id
+      @record_type = rec.class.name
+      @record = rec
     end
 
     def record_id
-      @record_id ||= record.id
+      @record_id ||= record ? record.id : nil
+    end
+
+    def record_id=(id)
+      @record = nil
+      @record_id = id
     end
 
     def record_type
-      @record_type ||= record.class.name
+      @record_type ||= record ? record.class.name : nil
+    end
+
+    def record_type=(type)
+      @record = nil
+      @record_type = type.to_s
     end
 
     def config
@@ -32,6 +60,10 @@ module ApplicationRecordLogger
 
     def record_klass
       record_type.constantize
+    end
+
+    def find_record
+      record_klass.find_by(id: record_id) if record_type && record_id
     end
   end
 end