+++ test/dummy/app/models/application_record.rb
@@ -1,5 +1,3 @@
 class ApplicationRecord < ActiveRecord::Base
   self.abstract_class = true
-  include ApplicationRecordLogger
-  set_logging_callbacks!
 end